refactor: ChatsとChatでディレクトリを分けるためにChatsディレクトリを作成した
