+++ src/resources/ts/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import { Login } from '@/ts/Login/Login';
-import { Chats } from '@/ts/Chat/Chats';
+import { Chats } from '@/ts/Chats/Chats';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { Route } from '@/ts/layout/Route';
 import { createTheme, ThemeProvider } from '@mui/material';
